refactor: change the design of challenges to improve testability
The design of the challenge interface that takes parsedContent as an input was not convenient for testing purposes.
Because of this, I had to potentially setup all the objects parsed which forced us to write much more code and was less readable.

Plain text is actually a better things that we want to pass to test our challenge, so we switched from an external dependency to parser to an internal one.

If we migrate the 2 first challenges of 2023, the ChallengeFacade won't make sense anymore (maybe, we could directly call a Factory in the command?)